Fix 3rd party extensions order and add test to avoid regression
